+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added:
   * Heretic skins
+  * Void Reaver skins
 
 * Changed:
   * Made Artificer cloth visible when using different model skin